Make examples into valid Arduino sketches
The Arduino IDE requires that a sketch:

- Be in a folder that contains only basic letters, numbers, dash, and underscore. Spaces and + are not allowed.
- Contain a .ino file that matches the folder name

Previously, the example sketches of this library did not meet either of those requirements, and thus the Arduino IDE did not recognize them.

Converting the examples to valid sketches makes the examples available via the Arduino IDE's File > Examples > EasyBlynk8266 menu after installing the library.